move apply page v2 into pages dir
