+++ src/pages/ApplyPageV2/TeamApplicationCard/style.css
